fix!: move `fetch()` to tests so they are the only ones with delegates.
This is a first step towards creating a multi-step `fetch()` which can
also be used in the `gix` crate.
