+++ gix-protocol/src/fetch/mod.rs
@@ -1,12 +1,6 @@
 mod arguments;
 pub use arguments::Arguments;
 
-///
-pub mod delegate;
-#[cfg(any(feature = "async-client", feature = "blocking-client"))]
-pub use delegate::Delegate;
-pub use delegate::{Action, DelegateBlocking};
-
 mod error;
 pub use error::Error;
 ///
@@ -15,6 +9,3 @@ pub use response::Response;
 
 mod handshake;
 pub use handshake::upload_pack as handshake;
-
-#[cfg(test)]
-mod tests;
