+++ gix-protocol/src/lib.rs
@@ -35,11 +35,6 @@ pub use maybe_async;
 #[cfg(any(feature = "blocking-client", feature = "async-client"))]
 pub mod fetch;
 
-#[cfg(any(feature = "blocking-client", feature = "async-client"))]
-mod fetch_fn;
-#[cfg(any(feature = "blocking-client", feature = "async-client"))]
-pub use fetch_fn::{fetch, FetchConnection};
-
 mod remote_progress;
 pub use remote_progress::RemoteProgress;
