+++ gix-protocol/tests/protocol/fetch/v2.rs
@@ -1,9 +1,11 @@
 use bstr::ByteSlice;
 use gix_features::progress;
-use gix_protocol::{fetch, handshake, ls_refs, FetchConnection};
+use gix_protocol::{fetch, handshake, ls_refs};
 use gix_transport::Protocol;
 
-use crate::fetch::{helper_unused, oid, transport, CloneDelegate, CloneRefInWantDelegate, LsRemoteDelegate};
+use crate::fetch::{
+    _impl::FetchConnection, helper_unused, oid, transport, CloneDelegate, CloneRefInWantDelegate, LsRemoteDelegate,
+};
 
 #[maybe_async::test(feature = "blocking-client", async(feature = "async-client", async_std::test))]
 async fn clone_abort_prep() -> crate::Result {
@@ -19,7 +21,7 @@ async fn clone_abort_prep() -> crate::Result {
         gix_transport::client::git::ConnectMode::Daemon,
     );
     let agent = "agent";
-    let err = gix_protocol::fetch(
+    let err = crate::fetch(
         &mut transport,
         &mut dlg,
         helper_unused,
@@ -66,7 +68,7 @@ async fn ls_remote() -> crate::Result {
         gix_transport::client::git::ConnectMode::Daemon,
     );
     let agent = "agent";
-    gix_protocol::fetch(
+    crate::fetch(
         &mut transport,
         &mut delegate,
         helper_unused,
@@ -122,7 +124,7 @@ async fn ls_remote_abort_in_prep_ls_refs() -> crate::Result {
         Protocol::V2,
         gix_transport::client::git::ConnectMode::Daemon,
     );
-    let err = gix_protocol::fetch(
+    let err = crate::fetch(
         &mut transport,
         &mut delegate,
         helper_unused,
@@ -164,7 +166,7 @@ async fn ref_in_want() -> crate::Result {
     );
 
     let agent = "agent";
-    gix_protocol::fetch(
+    crate::fetch(
         &mut transport,
         &mut delegate,
         helper_unused,
